- Tests moved as integration-test for Maven - New schema tests
