+++ src/test/java/com/jaeksoft/searchlib/test/legacy/SearchTest.java
@@ -22,7 +22,7 @@
  *  If not, see <http://www.gnu.org/licenses/>.
  **/
 
-package com.jaeksoft.searchlib.test;
+package com.jaeksoft.searchlib.test.legacy;
 
 import java.io.IOException;
 import java.util.ArrayList;
@@ -39,6 +39,7 @@ import org.junit.FixMethodOrder;
 import org.junit.Test;
 import org.junit.runners.MethodSorters;
 import org.xml.sax.SAXException;
+
 
 @FixMethodOrder(MethodSorters.NAME_ASCENDING)
 public class SearchTest extends TestCase {
