+++ src/test/java/com/jaeksoft/searchlib/test/legacy/WebCrawlerTest.java
@@ -1,7 +1,7 @@
 /**   
  * License Agreement for OpenSearchServer
  *
- * Copyright (C) 2012 Emmanuel Keller / Jaeksoft
+ * Copyright (C) 2012 - 2013 Emmanuel Keller / Jaeksoft
  * 
  * http://www.open-search-server.com
  * 
@@ -21,7 +21,8 @@
  *  along with OpenSearchServer. 
  *  If not, see <http://www.gnu.org/licenses/>.
  **/
-package com.jaeksoft.searchlib.test;
+
+package com.jaeksoft.searchlib.test.legacy;
 
 import java.io.IOException;
 import java.util.ArrayList;
@@ -34,27 +35,51 @@ import junit.framework.TestCase;
 
 import org.apache.http.NameValuePair;
 import org.apache.http.client.methods.HttpPost;
+import org.junit.FixMethodOrder;
 import org.junit.Test;
+import org.junit.runners.MethodSorters;
 import org.xml.sax.SAXException;
 
-public class OptimizeTest extends TestCase {
+
+/**
+ * @author Ayyathurai N Naveen
+ * 
+ */
+@FixMethodOrder(MethodSorters.NAME_ASCENDING)
+public class WebCrawlerTest extends TestCase {
 	private CommonTestCase commomTestCase = null;
 
-	public OptimizeTest(String name) {
+	public WebCrawlerTest(String name) {
 		super(name);
 		commomTestCase = new CommonTestCase();
 	}
 
-	@Test
-	public void testOptimizeIndex() throws IllegalStateException, IOException,
-			XPathExpressionException, SAXException,
+	public String startStopCrawler(String action) throws IllegalStateException,
+			IOException, XPathExpressionException, SAXException,
 			ParserConfigurationException {
 		List<NameValuePair> namedValuePairs = new ArrayList<NameValuePair>();
+		namedValuePairs.add(commomTestCase.getNameValuePair("action", action));
 		HttpPost httpPost = commomTestCase.queryInstance(namedValuePairs,
-				CommonTestCase.OPTIMIZE_API, true);
-		String response = commomTestCase.getHttpResponse(httpPost,
-				"response/entry[@key='Status']");
-		assertEquals("OK", response);
+				CommonTestCase.WEBCRAWLER_API, true);
+		return commomTestCase.getHttpResponse(httpPost,
+				"response/entry[@key='Info']");
+
+	}
+
+	@Test
+	public void testAStartCrawler() throws IllegalStateException,
+			XPathExpressionException, IOException, SAXException,
+			ParserConfigurationException {
+		String response = startStopCrawler("start");
+		assertEquals("STARTED", response);
+	}
+
+	@Test
+	public void testBStopCrawler() throws IllegalStateException,
+			XPathExpressionException, IOException, SAXException,
+			ParserConfigurationException {
+		String response = startStopCrawler("stop");
+		assertEquals("STOPPED", response);
 	}
 
 }
