+++ src/test/resources/com/jaeksoft/searchlib/test/rest/documents.xml
